requirements.txt
Django==3.0.2
Pillow==7.0.0
psycopg2-binary==2.8.4
pre-commit==2.1.0
django-filter